Shorten buffer

+++ src/source.rs
@@ -348,7 +348,7 @@ impl SoundSource {
             }
             // Fast-path for common case when there is no resampling and no pitch change.
             let from = self.buf_read_pos as usize;
-            let buffer_len = buffer.samples.len() / buffer.channel_count;
+            let buffer_len = buffer.samples.len() / usize::from(buffer.channel_count);
             let rendered = (buffer_len - from).min(amount);
             if buffer.channel_count == 2 {
                 for i in from..from + rendered {
@@ -405,7 +405,7 @@ impl SoundSource {
         let rel_step = step as f32;
         // We skip one last element because the hot loop resampling between current and next
         // element. Last elements are appended after the hot loop.
-        let buffer_last = buffer.samples.len() / buffer.channel_count - 1;
+        let buffer_last = buffer.samples.len() / usize::from(buffer.channel_count) - 1;
         if buffer.channel_count == 2 {
             while rendered < amount {
                 let (idx, w) = {
